Add daemon to monitor log STH freshness.
This commit adds the initial `ct-woodpecker` command. Once invoked this
command starts a daemon process that offers a HTTP server with
a Prometheus `/metrics` endpoint. The daemon remains resident after
startup to monitor configured certificate transparency logs, publishing
observations as metrics data.

Each monitored log will have its STH fetched periodically. The latency
of the fetch operation is published as a Prometheus histogram
observation. Any errors in fetching the STH will increment a Prometheus
counter. If the STH fetch succeeds and the signature validates
successfully then the STH timestamp is published as a Prometheus gauge
observation. This allows writing an alert to ensure that the monitored
STH timestamps are always increasing as expected.

This is *not* sufficient for a full blown CT monitor but it's a start
:-)

TODO: Unit tests!