Add unit tests for existing code.
This commit adds unit tests for the `ct-woodpecker/main.go` code.
Primarily this tests the configuration validation and loading.

This commit also adds unit tests for the monitor package. Primarily this
tests the `observeSTH` function and the `New` constructor.

Finally, unit tests are run in `-race` build mode in CI as part of
`.travis.yml`.