integration: allow for some jitter in woodpeckerRun.
Prior to this commit `woodpeckerRun` let a woodpecker instance under
test run for _exactly_ the time required to do a fixed number of
monitoring iterations based on the configured STH fetch interval. In
practice we can't expect the monitors to complete in exactly that amount
of time every time. This commit oversleeps for 20ms to ensure that minor
scheduling differences in the monitor goroutines don't cause the test to
fail. This resolves the failures where the monitor was killed early and
did not complete its expected work. Running integration tests in a tight
loop locally with this fix applied showed no failures.

Unfortunately the above is not enough to fix all CI failures and so this
commit also loosens the `sth_latency_count` and expected number of STH
fetch tests to accept over-checking the STH (e.g. because the test ran
longer than expected before killing the woodpecker instance).

In practice the combination of these two fixes results in 10/10 test
builds passing. Prior to these fixes only ~6-7/10 of builds pass.