integration: allow for some jitter in woodpeckerRun.
Prior to this commit `woodpeckerRun` let a woodpecker instance under
test run for _exactly_ the time required to do a fixed number of
monitoring iterations based on the configured STH fetch interval. In
practice we can't expect the monitors to complete in exactly that amount
of time every time. This commit oversleeps for 20ms to ensure that minor
scheduling differences in the monitor goroutines don't cause the test to
fail.