monitor: periodically submit certs from special root.
This commit adds a new feature to ct-woodpecker/monitors to periodically
issue throw-away test certificates with a special purpose signing
key/root certificate. To-be-monitored logs can add this root to their
accepted root configuration to allow ct-woodpecker to effectively
monitor cert submission without needing to involve the complexity of
issuing certificates from an established trusted root. A small command
line tool `ct-woodpecker-genissuer` is provided to spit out an ECDSA
private key and corresponding root certificate that can be used for this
purpose. Log operators will want to ensure the private key for the dummy
root added to the log's accepted roots is not made available to
untrusted parties as it may enable log-spamming.

Each log configured with `SubmitCert: true` has a testing root issued
certificate submitted to its add-chain endpoint on a fixed schedule.
Latency of the submissions as well as a success/failure count are
exposed via Prometheus metrics for monitoring purposes. Presently only
final certificates (not precerts) are submitted. If the returned SCT is
invalid (e.g. has a bad signature, wrong log ID) or if it is outside of
a fixed range of freshness (>10m into the past/future) then the
submission is considered an error result.