Process STH fetching and (pre)cert submission async.
Prior to this commit each log's monitor's `STHFetcher` or
`CertSubmitter` would block on the STH fetch or cert submission
completing or timing out, sleep for the cycle duration, and then repeat.

The problem with this approach is that if a log has high latency the
monitoring schedule will gradually skew and we end up performing less
STH fetches/cert submissions than intended.

This commit updates the `certSubmitter` and `sthFetcher` to perform the
STH fetching and cert submission in separate Go routines. This ensures
that every cycle tries to perform its work even if the work ends up
lagging.

To test this works as intended an integration tests configuring
a woodpecker to operate against two CT logs, one that is very slow and
one that is very fast, are included. These tests run the woodpecker for
the correct overall runtime based on the configured cycle time. The
standard out emitted by the woodpecker is checked to ensure the correct
number of STH fetches & cert submissions were attempted. Prior to
landing the changes in this branch these tests failed: only 1 operation
was performed against the slow log because subsequent cycles blocked.
With the changes in this branch applied the test passes and the correct
number of operations are attempted against the slow log.

Along the way I exposed the timeout for both STH fetching and cert
submission back through the configuration layers to allow setting
a sensible timeout for both operations per-log.

Similarly, I updated the `test/config.json` example - it had fallen out
of date and would not load.