Integration tests: relax testing requirements.
Prior to this commit the ct-woodpecker integration tests were a bit
flaky. The root cause is our usage of elapsed ct-woodpecker runtime as
a proxy for work completed. This is hard to do with accuracy and so (as
one example) the time required to fetch metrics output before stopping
the ct-woodpecker instance could result in another sth fetch/cert submit
cycle happening before the test completes, causing miscounts of expected
operations.

This commit changes the tests to just ensure a minimum of 1 cycles worth
of work (plus the startup cycle) occurs during testing. This is Good
Enough and will save us from considerable headaches tweaking padding
timeouts or overcomplicating the design of ct-woodpecker to expose cycle
information back to the test rig.

This commit also adds clean shutdown code to the monitors. This prevents
ghost monitor goroutines from sticking around when the parent
ct-woodpecker is stopped.