STH Fetcher: Verify consistency between STHs.
After fetching more than one STH the monitor will now verify the
consistency between the current STH and the previously observed STH.

Errors are tracked in the Prometheus `sth_inconsistencies` counter vec
sliced by the log URI and the type of inconsistency (bad STHs, failure
to get a consistency proof, failure to validate the consistency proof).

The latency of fetching a STH consistency proof is tracked in the
Prometheus `sth_proof_latency` histogram.