inclusionChecker: set `MaxGetEntries` from inclusionOptions. (#56)
The `monitor.New` constructor that created the `inclusionChecker` struct
was not wiring through the `opts.InclusionOpts.MaxGetEntries` setting
from the provided inclusion checker configuration. This results in an
OOM condition in `getEntries` when no clipping on the end range is done
and an entire log's worth of entries are built up in a single slice.