go mod: init module.
Notably this pins Trillian to d0c8d5f - the tip of master at the time
of writing. We're using changes that aren't in the latest v1.2.1
tag.

Similarly, Trillian at tip of master requires us to pin golang/protobuf
at 1d3f30b - the tip of master for that project at the time of writing.
Trillian is using changes in this project that aren't in the latest
v1.2.0 tag.