docker-compose: ct-test-srv, ct-woodpecker, and prometheus.
This adds docker support for building images for `ct-test-srv`,
`ct-woodpecker`, and `ct-malformed`.

The provided `docker-compose.yml` can be used to create a `ct-test-srv`
container providing two certificate transparency logs `log-one` and
`log-two`. A `ct-woodpecker` container is configured to monitor both
logs storing the data the mounted sqlite3 db. Lastly a `prometheus`
container is configured to scrape the `ct-woodpecker` metrics.