End-to-end docker compose setup w/ Prometheus/Alertmanager.
This adds a Prometheus and an Alertmanager container to the docker
compose environment. This environment is configured so that the
ct-woodpecker container is being scraped and alerts dispatched via the
alertmanager container. The alertmanager container uses a webhook that
sends the alerts to a `ct-test-srv` instance for logging to stdout.

To allow testing the rules the ct-test-srv now has a way to add mock
responses for any path. This can be used (for e.g.) to configure one of
the logs to return a 500 error for `/ct/v1/add-chain` for a period of
time.