Fix handling of notBefore/notAfter in issued certs
If windowStart/windowEnd were defined, an issued
cert would be created with notBefore/notAfter
set close to those same values. This would cause
problems issuing to a shard that wouldn't accept
certs dated in the future (which is Trillian's
behavior with "reject_expired: true").

Now certs issued for sharded logs, notBefore is
set to now and notAfter set to ~windowEnd, except
when windowEnd is in the past. In that case the
cert is backdated 90 days from windowEnd.