Update trillian and ct-go dependencies
The biggest changes here are that trillian stopped signing things,
leaving that responsibility to ctfe or equivalent. So a lot of the stuff
related to having trillian sign things is gone.

Also the trillian protobuf APIs started sending raw bytes instead of
separate fields, so there are changes to consume those raw bytes.