Use 90 day test certs in temporal shard window (#128)
Change the behavior of the test certificate generation to always create a
90 day certificate, and make sure it falls in the temporal shard window if
one is defined. If issuing a certificate for a current temporal shard
(where the system time + 90d NotAfter lands in the temporal window),
then generate a certifcate based on the current time. If generating a
certificate for a past or future temporal shard, then generate a 90d
certificate that falls in that window.

In addition, fix failing CI with updates and lint fixes:
* Update golangci-lint
* Add the ability to run tests and lint manually in GitHub Actions
* Add lint exclusions
* Fix misspelling
* Remove duplicate tls import

Fixes #127